LayoutTests/imported/w3c/web-platform-tests/storage-access-api/requestStorageAccess-cross-origin-iframe-navigation-relax.sub.https.window-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: SyntaxError: The string did not match the expected pattern.

Harness Error (FAIL), message = Unhandled rejection: The string did not match the expected pattern.

TIMEOUT Same-site-initiated same-origin navigations preserve storage access Test timed out
NOTRUN Same-site-initiated cross-origin navigations do not preserve storage access

Harness Error (FAIL), message = Unhandled rejection: The string did not match the expected pattern.

TIMEOUT Same-site-initiated same-origin navigations preserve storage access Test timed out
NOTRUN Same-site-initiated cross-origin navigations do not preserve storage access
